Move back to storing all images directly in context
