+++ jingle_sleigh/src/ffi/mod.rs
@@ -3,7 +3,6 @@ pub(crate) mod context_ffi;
 pub(crate) mod image;
 pub(crate) mod instruction;
 pub(crate) mod opcode;
-pub(crate) mod sleigh_image;
 
 // Need to pull this in somewhere so that libz symbols are available
 // for the `sleigh` CPP code at link-time.
@@ -20,11 +19,9 @@ mod tests {
         let builder =
             SleighContextBuilder::load_ghidra_installation("/Applications/ghidra").unwrap();
 
-        let sleigh = builder
-            .build("x86:LE:64:default")
-            .unwrap();
-        let bin_image = sleigh.load_image(bytes.as_slice()).unwrap();
-        let _lib = bin_image.instruction_at(0).unwrap();
+        let mut sleigh = builder.build("x86:LE:64:default").unwrap();
+        sleigh.set_image(bytes.as_slice()).unwrap();
+        sleigh.instruction_at(0).unwrap();
     }
     #[test]
     fn test_callother_decode2() {
@@ -32,10 +29,8 @@ mod tests {
         let builder =
             SleighContextBuilder::load_ghidra_installation("/Applications/ghidra").unwrap();
 
-        let sleigh = builder
-            .build("x86:LE:64:default")
-            .unwrap();
-        let bin_image = sleigh.load_image(bytes.as_slice()).unwrap();
-        let _lib = bin_image.instruction_at(0).unwrap();
+        let mut sleigh = builder.build("x86:LE:64:default").unwrap();
+        sleigh.set_image(bytes.as_slice()).unwrap();
+        sleigh.instruction_at(0).unwrap();
     }
 }
